examples/requirements.txt
django-debug-toolbar==5.0.1
PyYAML==6.0.2
